requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
pybase62==0.6.0
validators==0.20.0